refactor: fix Checkout feature exports
export CheckoutPage instead of CheckoutForm
