+++ src/features/Checkout/index.tsx
@@ -1,1 +1,1 @@
-export * from "./CheckoutForm";
+export * from "./CheckoutPage";
